Defined wiki_search function, new data structure
Defined wiki search function. Expected input is a search term. Returns
a list of dictionaries

Instead of storing all of the results in 1 dictionary each result is in
its own dictionary and is appended to a list of results. This makes it
easier to print out the title, description, and url simultaneously