Wikipedia Search Module
created wiki.py

wiki.py constructs a custom search url be designating how many results,
the format of the page (must be json), and the query

2 functions:

1.
wiki_search appends the search term onto the url and loads the json
data.
wiki_search then reads the json data and stores it in 3 lists (title,
description, url).
wiki_search then takes each result and appends it to a list. each
element of the list is a dictionary containing the title, description
and url
wiki_search the returns this list of dictionaries

*thought*: I suppose I could split up wiki_search to create the url
then read the json data to support a stronger modular design

2. wiki_search_list
wiki_search_list calls wiki_search and converts the list of
dictionaries to a list of titles with their respective descriptions
it then returns said list