Implemented CGI
- Using CGI to run search.py

- the client will enter a search term into the form then it uses the
HTTP GET method to pass the query to the web server (this also appends
the query to the URL)

- search now prints results with paragraph tags
* right now the combo_search function sometimes doesn’t return the
expected data structure (a list) - this bug messes up the output when
the results are printed because the for loop is set up for a list

- **also another BIG problem** the python library I’m using to access
the Twitter API, Tweepy, isn’t working on the server. I need to do some
research why. I pip installed within the cgi bin folder and even
changed the permissions of all the files to 755. I need to figure this
out or access Twitter another way.
-therefore I have commented out the calls to twit.py in order to get a
skeleton web application running

I plan on:
Fixing the bug with combo_search
fixing the twitter API
posting the results to the same search page using javascript (?)